+++ packages/core/demo/demo-data/line.ts
@@ -1,5 +1,4 @@
 import { colors } from "./colors";
-import { getTheme } from "./themes";
 import { Tools } from "../../src/tools";
 
 export const lineTimeSeriesData = {
@@ -7,7 +6,7 @@ export const lineTimeSeriesData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			backgroundColors: [colors[0]],
+			fillColors: [colors[0]],
 			data: [
 				{
 					date: new Date(2019, 0, 1),
@@ -33,7 +32,7 @@ export const lineTimeSeriesData = {
 		},
 		{
 			label: "Dataset 2",
-			backgroundColors: [colors[1]],
+			fillColors: [colors[1]],
 			data: [
 				{
 					date: new Date(2019, 0, 2),
@@ -59,7 +58,7 @@ export const lineTimeSeriesData = {
 		},
 		{
 			label: "Dataset 3",
-			backgroundColors: [colors[2]],
+			fillColors: [colors[2]],
 			data: [
 				{
 					date: new Date(2019, 0, 1),
@@ -85,7 +84,7 @@ export const lineTimeSeriesData = {
 		},
 		{
 			label: "Dataset 4",
-			backgroundColors: [colors[3]],
+			fillColors: [colors[3]],
 			data: [
 				{
 					date: new Date(2019, 0, 2),
@@ -119,12 +118,11 @@ export const lineTimeSeriesOptions = {
 			secondary: true
 		},
 		bottom: {
-			type: "time",
+			scaleType: "time",
 			primary: true
 		}
 	},
-	curve: "curveMonotoneX",
-	theme: getTheme()
+	curve: "curveMonotoneX"
 };
 
 export const lineData = {
@@ -132,7 +130,7 @@ export const lineData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			backgroundColors: [colors[0]],
+			fillColors: [colors[0]],
 			data: [
 				32100,
 				23500,
@@ -143,7 +141,7 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 2",
-			backgroundColors: [colors[1]],
+			fillColors: [colors[1]],
 			data: [
 				34200,
 				53200,
@@ -154,7 +152,7 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 3 long name",
-			backgroundColors: [colors[2]],
+			fillColors: [colors[2]],
 			data: [
 				41200,
 				23400,
@@ -165,7 +163,7 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 4 long name",
-			backgroundColors: [colors[3]],
+			fillColors: [colors[3]],
 			data: [
 				22000,
 				1200,
@@ -176,7 +174,7 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 5 long name",
-			backgroundColors: [colors[4]],
+			fillColors: [colors[4]],
 			data: [
 				2412,
 				30000,
@@ -187,7 +185,7 @@ export const lineData = {
 		},
 		{
 			label: "Dataset 6 long name",
-			backgroundColors: [colors[5]],
+			fillColors: [colors[5]],
 			data: [
 				0,
 				20000,
@@ -204,14 +202,13 @@ export const lineOptions = {
 	axes: {
 		bottom: {
 			title: "2018 Annual Sales Figures",
-			type: "labels",
+			scaleType: "labels",
 			secondary: true
 		},
 		left: {
 			primary: true
 		}
-	},
-	theme: getTheme()
+	}
 };
 
 // Step
@@ -235,7 +232,7 @@ export const scatterData = {
 	datasets: [
 		{
 			label: "Dataset 1",
-			backgroundColors: [colors[0]],
+			fillColors: [colors[0]],
 			data: [
 				32100,
 				23500,
@@ -246,7 +243,7 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 2",
-			backgroundColors: [colors[1]],
+			fillColors: [colors[1]],
 			data: [
 				34200,
 				53200,
@@ -257,7 +254,7 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 3 long name",
-			backgroundColors: [colors[2]],
+			fillColors: [colors[2]],
 			data: [
 				41200,
 				23400,
@@ -268,7 +265,7 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 4 long name",
-			backgroundColors: [colors[3]],
+			fillColors: [colors[3]],
 			data: [
 				22000,
 				1200,
@@ -279,7 +276,7 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 5 long name",
-			backgroundColors: [colors[4]],
+			fillColors: [colors[4]],
 			data: [
 				2412,
 				30000,
@@ -290,7 +287,7 @@ export const scatterData = {
 		},
 		{
 			label: "Dataset 6 long name",
-			backgroundColors: [colors[5]],
+			fillColors: [colors[5]],
 			data: [
 				0,
 				20000,
@@ -307,14 +304,13 @@ export const scatterOptions = {
 	axes: {
 		bottom: {
 			title: "2018 Annual Sales Figures",
-			type: "labels",
+			scaleType: "labels",
 			secondary: true
 		},
 		left: {
 			primary: true
 		}
-	},
-	theme: getTheme()
+	}
 };
 
 export const scatterTimeSeriesData = lineTimeSeriesData;
@@ -324,12 +320,11 @@ export const scatterTimeSeriesOptions = {
 	axes: {
 		bottom: {
 			title: "2019 Annual Sales Figures",
-			type: "time",
+			scaleType: "time",
 			secondary: true
 		},
 		left: {
 			primary: true
 		}
-	},
-	theme: getTheme()
+	}
 };