+++ packages/core/src/components/axes/axis.ts
@@ -8,8 +8,8 @@ import { DOMUtils } from "../../services";
 // D3 Imports
 import { scaleBand, scaleLinear, scaleTime, scaleLog, scaleOrdinal } from "d3-scale";
 import { axisBottom, axisLeft, axisRight, axisTop } from "d3-axis";
-import { min, max, extent } from "d3-array";
-import { select } from "d3-selection";
+import { min, extent } from "d3-array";
+import { timeFormatDefaultLocale } from "d3-time-format";
 
 export class Axis extends Component {
 	type = "axes";
@@ -31,14 +31,15 @@ export class Axis extends Component {
 
 	createOrGetScale() {
 		const { position } = this.configs;
-		const scaleOptions = Tools.getProperty(this.model.getOptions(), "axes", position);
+		const axisOptions = Tools.getProperty(this.model.getOptions(), "axes", position);
+		this.scaleType = (axisOptions && axisOptions.scaleType) ? axisOptions.scaleType : ScaleTypes.LINEAR;
 
 		let scaleFunction;
-		if (scaleOptions && scaleOptions.type === ScaleTypes.TIME) {
+		if (this.scaleType === ScaleTypes.TIME) {
 			scaleFunction = scaleTime();
-		} else if (scaleOptions && scaleOptions.type === ScaleTypes.LOG) {
-			scaleFunction = scaleLog().base(scaleOptions.base || 10);
-		} else if (scaleOptions && scaleOptions.type === ScaleTypes.LABELS) {
+		} else if (this.scaleType === ScaleTypes.LOG) {
+			scaleFunction = scaleLog().base(axisOptions.base || 10);
+		} else if (this.scaleType === ScaleTypes.LABELS) {
 			scaleFunction = scaleBand();
 		} else {
 			scaleFunction = scaleLinear();
@@ -50,12 +51,12 @@ export class Axis extends Component {
 				[position]: this
 			};
 
-			if (scaleOptions) {
-				if (scaleOptions.primary === true) {
+			if (axisOptions) {
+				if (axisOptions.primary === true) {
 					modelUpdates[AxisTypes.PRIMARY] = this;
 				}
 
-				if (scaleOptions.secondary === true) {
+				if (axisOptions.secondary === true) {
 					modelUpdates[AxisTypes.SECONDARY] = this;
 				}
 			}
@@ -64,7 +65,6 @@ export class Axis extends Component {
 		}
 
 		this.scale = scaleFunction;
-		this.scaleType = (scaleOptions && scaleOptions.type) ? scaleOptions.type : ScaleTypes.LINEAR;
 
 		return scaleFunction;
 	}
@@ -76,12 +76,12 @@ export class Axis extends Component {
 	getScaleDomain() {
 		const options = this.model.getOptions();
 		const { position } = this.configs;
-		const scaleOptions = Tools.getProperty(options, "axes", position);
+		const axisOptions = Tools.getProperty(options, "axes", position);
 
 		const { datasets, labels } = this.model.getDisplayData();
 
 		// If scale is a LABELS scale, return some labels as the domain
-		if (scaleOptions && scaleOptions.type === ScaleTypes.LABELS) {
+		if (axisOptions && axisOptions.scaleType === ScaleTypes.LABELS) {
 			if (labels) {
 				return labels;
 			} else {
@@ -92,7 +92,7 @@ export class Axis extends Component {
 		// Get the extent of the domain
 		let domain;
 		// If the scale is stacked
-		if (scaleOptions.stacked) {
+		if (axisOptions.stacked) {
 			domain = extent(
 				labels.reduce((m, label: any, i) => {
 					const correspondingValues = datasets.map(dataset => {
@@ -111,7 +111,7 @@ export class Axis extends Component {
 			// Get all the chart's data values in a flat array
 			let allDataValues = datasets.reduce((dataValues, dataset: any) => {
 				dataset.data.forEach((datum: any) => {
-					if (scaleOptions.type === ScaleTypes.TIME) {
+					if (axisOptions.scaleType === ScaleTypes.TIME) {
 						dataValues = dataValues.concat(datum.date);
 					} else {
 						dataValues = dataValues.concat(isNaN(datum) ? datum.value : datum);
@@ -121,14 +121,14 @@ export class Axis extends Component {
 				return dataValues;
 			}, []);
 
-			if (scaleOptions.type !== ScaleTypes.TIME) {
+			if (axisOptions.scaleType !== ScaleTypes.TIME) {
 				allDataValues = allDataValues.concat(0);
 			}
 
 			domain = extent(allDataValues);
 		}
 
-		if (scaleOptions.type === ScaleTypes.TIME) {
+		if (axisOptions.scaleType === ScaleTypes.TIME) {
 			if (Tools.getProperty(options, "timeScale", "addSpaceOnEdges")) {
 				// TODO - Need to account for non-day incrementals as well
 				const [startDate, endDate] = domain;
@@ -167,7 +167,7 @@ export class Axis extends Component {
 		// Grab the scale off of the model, and initialize if it doesn't exist
 		const scale = this.createOrGetScale().domain(this.getScaleDomain());
 
-		if (axisOptions && axisOptions.type === ScaleTypes.LABELS) {
+		if (this.scaleType === ScaleTypes.LABELS) {
 			scale.rangeRound([startPosition, endPosition]);
 		} else {
 			scale.range([startPosition, endPosition]);
@@ -190,9 +190,18 @@ export class Axis extends Component {
 				break;
 		}
 
+		// Set the date/time locale
+		if (this.scaleType === ScaleTypes.TIME) {
+			const timeLocale = Tools.getProperty(options, "locale", "time");
+			if (timeLocale) {
+				timeFormatDefaultLocale(timeLocale);
+			}
+		}
+
 		// Initialize axis object
 		const axis = axisFunction(scale)
-			.tickSizeOuter(0);
+			.tickSizeOuter(0)
+			.tickFormat(Tools.getProperty(axisOptions, "ticks", "formatter"));
 
 		if (scale.ticks) {
 			const numberOfTicks = 7;